Move fmt_rule() and fmt_rust_function() to testutil.engine.util
They're useful to at least 3 files. They might be useful to other files too.

Also, update the docstring per Danny's recommendation.
