+++ tests/python/pants_test/engine/BUILD
@@ -111,7 +111,6 @@ python_tests(
   dependencies=[
     '3rdparty/python:dataclasses',
     ':scheduler_test_base',
-    ':rules',
     'src/python/pants/base:cmd_line_spec_parser',
     'src/python/pants/build_graph',
     'src/python/pants/engine:nodes',
@@ -184,6 +183,7 @@ python_tests(
     'src/python/pants/engine:rules',
     'src/python/pants/engine:selectors',
     'tests/python/pants_test/engine/examples:parsers',
+    'src/python/pants/testutil/engine:util',
     'src/python/pants/testutil/subsystem',
     'src/python/pants/testutil:test_base',
   ],
@@ -197,12 +197,12 @@ python_tests(
   dependencies=[
     '3rdparty/python:dataclasses',
     ':util',
-    ':rules',
     'src/python/pants/base:cmd_line_spec_parser',
     'src/python/pants/build_graph',
     'src/python/pants/engine:scheduler',
     'tests/python/pants_test/engine/examples:scheduler_inputs',
     'src/python/pants/testutil:test_base',
+    'src/python/pants/testutil/engine:util',
   ],
   tags = {"partially_type_checked"},
 )
