+++ tests/python/pants_test/engine/test_engine.py
@@ -10,9 +10,13 @@ from pants.engine.rules import RootRule, rule
 from pants.engine.scheduler import ExecutionError
 from pants.engine.selectors import Get, MultiGet
 from pants.reporting.streaming_workunit_handler import StreamingWorkunitHandler
-from pants.testutil.engine.util import assert_equal_with_printing, remove_locations_from_traceback
+from pants.testutil.engine.util import (
+  assert_equal_with_printing,
+  fmt_rule,
+  fmt_rust_function,
+  remove_locations_from_traceback,
+)
 from pants_test.engine.scheduler_test_base import SchedulerTestBase
-from pants_test.engine.test_rules import fmt_rule, fmt_task_func
 
 
 class A:
@@ -181,7 +185,7 @@ class EngineTest(unittest.TestCase, SchedulerTestBase):
     self.assert_equal_with_printing(dedent(f'''
       1 Exception encountered:
       Computing Select(<{__name__}.B object at 0xEEEEEEEEE>, A)
-        Computing Task({fmt_task_func(nested_raise)}(), <{__name__}.B object at 0xEEEEEEEEE>, A, true)
+        Computing Task({fmt_rust_function(nested_raise)}(), <{__name__}.B object at 0xEEEEEEEEE>, A, true)
           Throw(An exception for B)
             Traceback (most recent call last):
               File LOCATION-INFO, in call
