+++ tests/python/pants_test/engine/test_scheduler.py
@@ -14,9 +14,12 @@ from pants.engine.objects import union
 from pants.engine.rules import RootRule, UnionRule, rule
 from pants.engine.scheduler import ExecutionError, SchedulerSession
 from pants.engine.selectors import Get, Params
-from pants.testutil.engine.util import assert_equal_with_printing, remove_locations_from_traceback
+from pants.testutil.engine.util import (
+  assert_equal_with_printing,
+  fmt_rust_function,
+  remove_locations_from_traceback,
+)
 from pants.testutil.test_base import TestBase
-from pants_test.engine.test_rules import fmt_task_func
 
 
 @dataclass(frozen=True)
@@ -263,7 +266,7 @@ class SchedulerWithNestedRaiseTest(TestBase):
     expected_msg = (
       "Exception: WithDeps(Inner(InnerEntry { params: {TypeCheckFailWrapper}, rule: Task(Task { "
       "product: A, clause: [Select { product: TypeCheckFailWrapper }], gets: [Get { product: A, "
-      f"subject: B }}], func: {fmt_task_func(a_typecheck_fail_test)}(), cacheable: true, display_info: "
+      f"subject: B }}], func: {fmt_rust_function(a_typecheck_fail_test)}(), cacheable: true, display_info: "
       "None }) })) did not declare a dependency on JustGet(Get { product: A, subject: A })"
     )
     with assert_execution_error(self, expected_msg):
@@ -351,7 +354,7 @@ class SchedulerWithNestedRaiseTest(TestBase):
       dedent(
         f'''\
         Computing Select(B(), A)
-          Computing Task({fmt_task_func(nested_raise)}(), B(), A, true)
+          Computing Task({fmt_rust_function(nested_raise)}(), B(), A, true)
             Throw(An exception for B)
               Traceback (most recent call last):
                 File LOCATION-INFO, in call
